+++ test/test229.js
@@ -160,28 +160,62 @@ assert.deepEqual(res,[ { even: 0, id: 2, value: 2, total: 2 },
 
     it("7. Select accumulated sum", function(done) {
 
-var res = alasql('SELECT [value]%2, a.id, a.[value], SUM(b.[value]) \
+alasql.fn.mod = function(a,b){return a%2 == b%2};
+
+var res = alasql('SELECT [value]%2 AS even, id, [value], SUM(b.[value]) AS total \
                     FROM   RunTotalTestData a, \
                             RunTotalTestData b \
-                    WHERE b.id      <= a.id \
-                          AND   b.[value]%2  = a.[value]%2 \
+                    WHERE a.[value]%2 = b.[value]%2 \
+                    AND b.id <= a.id \
                     GROUP BY a.[value]%2, a.id, a.[value] \
                     ORDER BY [value]%2, id;');
-console.log(res);
-// alasql('SELECT a.id, a.[value], (SELECT SUM(b.[value])  \
-//                                FROM RunTotalTestData b \
-//                                WHERE b.id <= a.id \
-//                                AND b.[value]%2 = a.[value]%2) \
-// FROM   RunTotalTestData a \
-// ORDER BY a.id;');
 
-//console.log(res);
-//    	var res = alasql('UPDATE one SET a = a + (SELECT MAX(b) FROM one)');
-//      console.log(res);
-//    	assert.deepEqual(res,[1.23, 2.345, 4.56]);
+assert.deepEqual(res,[ { even: 0, id: 2, value: 2, total: 2 },
+  { even: 0, id: 3, value: 4, total: 6 },
+  { even: 0, id: 6, value: 12, total: 18 },
+  { even: 0, id: 8, value: 16, total: 34 },
+  { even: 0, id: 9, value: 22, total: 56 },
+  { even: 0, id: 10, value: 42, total: 98 },
+  { even: 0, id: 12, value: 58, total: 156 },
+  { even: 0, id: 14, value: 60, total: 216 },
+  { even: 1, id: 1, value: 1, total: 1 },
+  { even: 1, id: 4, value: 7, total: 8 },
+  { even: 1, id: 5, value: 9, total: 17 },
+  { even: 1, id: 7, value: 13, total: 30 },
+  { even: 1, id: 11, value: 57, total: 87 },
+  { even: 1, id: 13, value: 59, total: 146 } ]);
 
       done();
     });
+//}
+
+
+// SQL NOT REALIZED YET
+    it("8. Over 1", function(done) {
+        var ast = alasql.parse('SELECT a.id, a.[value], SUM(a.[value]) OVER (ORDER BY a.id) \
+          FROM   RunTotalTestData a \
+          ORDER BY a.id;');
+//        console.log(ast.toString());
+        done();
+    });
+
+    it("8. Over 2", function(done) {
+        var ast = alasql.parse('SELECT a.id, a.[value], SUM(a.[value]) OVER (ORDER BY a.id) \
+            FROM   RunTotalTestData a \
+            WHERE a.[value] % 2 = 1 \
+            ORDER BY a.id;');
+//        console.log(ast.toString());
+        done();
+    });
+
+    it("9. Over 3", function(done) {
+        var ast = alasql.parse('SELECT a.value%2, a.id, a.[value], SUM(a.[value]) OVER (PARTITION BY a.[value]%2 ORDER BY a.id) \
+            FROM   RunTotalTestData a \
+            ORDER BY a.[value]%2, a.id;');
+//        console.log(ast.toString());
+        done();
+    });
+
 
     it("99. Drop database", function(done) {